requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
fastapi==0.120.1
fonttools==4.60.1
graphql-core==3.2.6
h11==0.16.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
lia-web==0.2.3
lightgbm==4.6.0
matplotlib==3.10.7
motor==3.7.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
strawberry-graphql==0.284.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unicodedata2==16.0.0
urllib3==2.5.0
uvicorn==0.38.0
xgboost==3.1.1
